+++ api/hydrophoneApi_test.go
@@ -136,10 +136,10 @@ func (m *testingShorelingMock) GetUser(userID, token string) (*schema.UserData, 
 func (m *testingShorelingMock) UpdateUser(userID string, userUpdate schema.UserUpdate, token string) error {
 	return nil
 }
-func (m *testingShorelingMock) GetUnverifiedUsers() ([]schema.UserData, error) {
+func (m *testingShorelingMock) GetUnverifiedUsers(token string) ([]schema.UserData, error) {
 	return []schema.UserData{}, nil
 }
-func (m *testingShorelingMock) DeleteUser(userID string) error {
+func (m *testingShorelingMock) DeleteUser(userID string, token string) error {
 	return nil
 }
 func (m *testingShorelingMock) CheckToken(chkToken string) *token.TokenData {